Improved descriptor methods; client better deal with timeout request than before

+++ src/main/scala/com/bwsw/tstreamstransactionserver/netty/client/ClientInitializer.scala
@@ -10,7 +10,7 @@ import io.netty.handler.codec.bytes.ByteArrayEncoder
 
 import scala.concurrent.{ExecutionContext, Promise => ScalaPromise}
 
-class ClientInitializer(reqIdToRep: ConcurrentHashMap[Integer, ScalaPromise[ThriftStruct]], client: Client, context: ExecutionContext) extends ChannelInitializer[SocketChannel] {
+class ClientInitializer(reqIdToRep: ConcurrentHashMap[Long, ScalaPromise[ThriftStruct]], client: Client, context: ExecutionContext) extends ChannelInitializer[SocketChannel] {
   override def initChannel(ch: SocketChannel): Unit = {
     ch.pipeline()
       .addLast(new ByteArrayEncoder())
